- Odds and ends  - Added the ability to define default storage slots per world
git-svn-id: svn://70.113.44.214/VanaMirror/branches/buis_vana_overhaul@3137 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/Drop.cpp
@@ -90,7 +90,7 @@ void Drop::showDrop(Player *player) {
 	DropsPacket::showDrop(player, this, DropsPacket::DropTypes::ShowExisting, false, Pos());
 }
 
-void Drop::takeDrop(Player *player, int32_t petId) {
+void Drop::takeDrop(Player *player, int64_t petId) {
 	Maps::getMap(m_mapId)->removeDrop(m_id);
 	if (petId == 0) {
 		DropsPacket::takeDrop(player, this);
