+++ src/ChannelServer/PlayerStorage.h
@@ -31,19 +31,19 @@ public:
 	PlayerStorage(Player *player);
 	~PlayerStorage();
 
-	void setSlots(int8_t slots);
+	void setSlots(uint8_t slots);
 	void addItem(Item *item);
-	void takeItem(int8_t slot);
+	void takeItem(uint8_t slot);
 	void setMesos(int32_t mesos) { m_mesos = mesos; }
 	void changeMesos(int32_t mesos);
 
-	int8_t getSlots() const { return m_slots; }
-	int8_t getNumItems() const { return m_items.size(); }
-	int8_t getNumItems(int8_t m_inv);
+	uint8_t getSlots() const { return m_slots; }
+	uint8_t getNumItems() const { return m_items.size(); }
+	uint8_t getNumItems(uint8_t m_inv);
 	int32_t getMesos() const { return m_mesos; }
-	bool isFull() const { return ((int8_t) m_items.size() == m_slots); }
-	Item * getItem(int8_t slot) const {
-		if (slot < (int8_t) m_items.size()) {
+	bool isFull() const { return (m_items.size() == m_slots); }
+	Item * getItem(uint8_t slot) const {
+		if (slot < m_items.size()) {
 			return m_items[slot];
 		}
 		return 0;
@@ -54,6 +54,6 @@ public:
 private:
 	Player *m_player;
 	vector<Item *> m_items;
-	int8_t m_slots;
+	uint8_t m_slots;
 	int32_t m_mesos;
 };
